wasm: upgrade m_index to long

+++ bsnes/wasm/wasminstance.hpp
@@ -93,7 +93,7 @@ public:
 public:
   WASMInterface* m_interface;
   const std::string m_key;
-  int m_index;
+  long m_index;
 
   std::shared_ptr<ZipArchive> m_za;
